changes and additions

+++ Assets/Scripts/Weapons/Childs/Hands.cs
@@ -5,13 +5,13 @@ using UnityEngine;
 public class Hands : Weapon
 {
     [SerializeField] private int attackDamage;
-    [SerializeField] private int attackSpeed;
+    [SerializeField] private float attackSpeed;
     private CharacterController player;
     private EnemyController enemy;
     private bool isAttacking = false;
     public override int AttackDamage => attackDamage;
 
-    public override int AttackSpeed => attackSpeed;
+    public override float AttackSpeed => attackSpeed;
 
     public override CharacterController Player => player;
     public override bool IsAttacking => isAttacking;
