+++ Assets/Scripts/Weapons/Parent/Weapon.cs
@@ -5,7 +5,7 @@ using UnityEngine;
 public abstract class Weapon : MonoBehaviour
 {
     public abstract int AttackDamage { get; }
-    public abstract int AttackSpeed { get; }
+    public abstract float AttackSpeed { get; }
     public abstract CharacterController Player { get; }
     public abstract bool IsAttacking { get; }
     public abstract void Attack(float agilityStat);
